# perf: minor `trace` improvements