small perf but just to get me started:

```
before:

10 passed in 74.92s (0:01:14)
10 passed in 74.68s

after:

10 passed in 73.99s (0:01:13)
10 passed in 73.17s (0:01:13)
```

must also use https://github.com/ApeWorX/ape-foundry/pull/121

* caching
* avoid some unnecessary computes